Main functions

FastSolversForWeightedTV.gradient_norm — Function

gradient_norm(p1, p2, n, h; weight=nothing, complex=false, gpu=false, options=exact_argmin())

Returns the regularization function associated to the weighted gradient mixed norm $g(\mathbf{u})=||\mathrm{A}\nabla\mathbf{u}||_{p1,p2}$.

The mixed norm is specified by p1 and p2. For p1=2, p2=1, and weight=nothing, for example, one gets conventional TV. Other supported options are p2=2 or p2=Inf.

The Cartesian grid geometry is determined by the grid size n and spacing h. For instance, in 3D, n=(64, 128, 256), h=(1f0, 2f0, 3f0).

The linear operator $A$ is specified via the keyword weight. Note that this weight must be initialized via the tools contained in the package AbstractLinearOperators (see Section Getting started for an example).

Complex or real inputs are handled via the keyword complex. Set complex=true for complex image input.

For sped-up computations, set gpu=true. Note that the input must also be a CuArray type, in this case.

For the evaluation of the associated proximal operator, one must specify a solver with the keyword options. Dedicated solvers are offered by the package AbstractProximableFunctions.jl (e.g. FISTA, see Section Getting started for some basic usage options).