Main functions

FastSolversForWeightedTV.gradient_norm — Function

gradient_norm(p1, p2, n, h; weight=nothing,
                            complex=false,
                            gpu=false,
                            options=exact_argmin())

Returns the regularization function associated to the weighted gradient mixed norm $g(\mathbf{u})=||\mathrm{A}\nabla\mathbf{u}||_{p1,p2}$.

The mixed norm is specified by p1 and p2. For p1=2, p2=1, and weight=nothing, for example, one gets conventional TV. Other supported options are p2=2 or p2=Inf.

The Cartesian grid geometry is determined by the grid size n and spacing h. For instance, in 3D, n=(64, 128, 256), h=(1f0, 2f0, 3f0).

The linear operator $A$ is specified via the keyword weight. Note that this weight must be initialized via the tools contained in the package AbstractLinearOperators (see Section Getting started for an example).

Complex or real inputs are handled via the keyword complex. Set complex=true for complex image input.

For sped-up computations, set gpu=true. Note that the input must also be a CuArray type, in this case.

For the evaluation of the associated proximal operator, one must specify a solver with the keyword options. Dedicated solvers are offered by the package AbstractProximableFunctions.jl (e.g. FISTA, see Section Getting started for some basic usage options).

Important utilities from ancillary packages

We list the most important utilities from AbstractProximableFunctions that are meant to be routinely combined with FastSolversForWeightedTV, such as tools to define iterative solver options, and functions to call the proximal/projection operators.

AbstractProximableFunctions.prox — Method

prox(y::AbstractArray{CT,N}, λ::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the proximal operator $\mathrm{prox}_{\lambda,g}(\mathbf{y})$.

AbstractProximableFunctions.prox — Method

prox(y::AbstractArray{CT,N}, λ::T, g::AbstractProximableFunction{CT,N}, options::AbstractArgminOptions) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the proximal operator with specified options.

See prox(y::AbstractArray{CT,N}, λ::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

AbstractProximableFunctions.proj — Method

proj(y::AbstractArray{CT,N}, ε::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the projection operator $\mathrm{proj}_{\lambda,g}(\mathbf{y})$.

AbstractProximableFunctions.proj — Method

proj(y::AbstractArray{CT,N}, ε::T, g::AbstractProximableFunction{CT,N}, options::AbstractArgminOptions) where {T<:Real,N,CT<:RealOrComplex{T}}

Computes the projection operator with specified options.

See proj(y::AbstractArray{CT,N}, ε::T, g::AbstractProximableFunction{CT,N}) where {T<:Real,N,CT<:RealOrComplex{T}}

AbstractProximableFunctions.exact_argmin — Function

exact_argmin()

Returns exact optimization options for the computation of proximal/projection operators. It might results in an error if the regularization functional does not implement analytically-defined proximal/projection operators (e.g. TV!).

AbstractProximableFunctions.FISTA_options — Function

FISTA_options(L; Nesterov=true,
                 reset_counter=nothing,
                 niter=nothing,
                 verbose=false,
                 fun_history=false)

Returns FISTA iterative solver options for the general optimization problem:

$\min_{\mathbf{x}}f(\mathbf{x})+g(\mathbf{x})$

where $g$ is a "proximable" function. It can be used to define options for the proximal or projection operators.

The parameter $L$ ideally should be chosen as $L\ge\mathrm{Lip}\nabla f$ (=Lipschitz constant of the gradient) and is problem specific.

Nesterov acceleration is set by Nesterov=true, while reset_counter is the number of iteration after which the Nesterov momentum is reset. The total number of iterations is determined by niter.

For debugging, set verbose=true and/or fun_history=true (the latter allows storing the history of $f(\mathbf{x})$, which can be retrieved by fun_history(options) after minimization).

Important note

When setting the options for proximal or projection operators, follow the recommendations for each specific proximal function on how to choose $L$. The underlying optimization may be based on algebraic reformulations of the optimization problem $\min_{\mathbf{x}}1/2||\mathbf{x}-\mathbf{y}||^2+\lambda{}g(\mathbf{x})$, so the assumption that $f(\mathbf{x})=1/2||\mathbf{x}-\mathbf{y}||^2$ is not generally correct. This note is relevant when computing the proximal operator of WeightedProximalFunction's for which the problem $\min_{\mathbf{x}}1/2||\mathbf{x}-\mathbf{y}||^2+\lambda{}g(A\mathbf{x})$ is transformed to $\min_{\mathbf{p}}1/2||\lambda A^*\mathbf{p}-\mathbf{y}||^2+\lambda g^*(\mathbf{p})$. The Lipschitz constant in this case should be $\lambda^2\rho(A)$. For convenience of use, however, the Lipschitz constant is expected to be just $\rho(A)$.